# packages/shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghar Khoj - Broker-Free Room Rental Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{userId}/{otherUserId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get conversation between two users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: otherUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
...
  /matches:
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a match (tenant shows interest in a room)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchInput"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/tenant/{tenantId}:
    get:
      operationId: getTenantMatches
      tags: [matches]
      summary: Get all matches for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
  /matches/owner/{ownerId}:
    get:
      operationId: getOwnerMatches
      tags: [matches]
      summary: Get all matches for an owner
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
  /matches/{id}/respond:
    patch:
      operationId: respondToMatch
      tags: [matches]
      summary: Owner or tenant responds to a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondMatchInput"
      responses:
        "200":
          description: Match updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        roomId:
          type: integer
        content:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, senderId, receiverId, content, isRead, createdAt]
    SendMessageInput:
      type: object
      properties:
        senderId:
          type: integer
        receiverId:
          type: integer
        roomId:
          type: integer
        content:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
      required: [senderId, receiverId]
    MessageListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required: [messages]
...
    Match:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        ownerId:
          type: integer
        roomId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled]
        tenantStatus:
          type: string
          enum: [pending, accepted, rejected]
        ownerStatus:
          type: string
          enum: [pending, accepted, rejected]
        matchScore:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tenantId, ownerId, roomId, status, tenantStatus, ownerStatus, createdAt, updatedAt]
    CreateMatchInput:
      type: object
      properties:
        tenantId:
          type: integer
        ownerId:
          type: integer
        roomId:
          type: integer
        matchScore:
          type: integer
      required: [tenantId, ownerId, roomId]
    RespondMatchInput:
      type: object
      properties:
        role:
          type: string
          enum: [tenant, owner]
        decision:
          type: string
          enum: [accepted, rejected]
      required: [role, decision]
    MatchListResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
      required: [matches]
...